refactor homepage sections
