+++ src/pages/Section2/cards.json
@@ -1,0 +1,26 @@
+[
+  {
+    "bannerUrl": "https://cosmos-x.oss-cn-hangzhou.aliyuncs.com/cardBanner1.svg",
+    "title": "Basics",
+    "intro": "The core language of frontend is JS. From here, you can learn not only JS/CSS/HTML, but also React, NodeJS, Webpack as well. In the following lessons, you'll get a healthy understanding of each of front-end.",
+    "moreUrl": "#basics"
+  },
+  {
+    "bannerUrl": "https://cosmos-x.oss-cn-hangzhou.aliyuncs.com/cardBanner2.svg",
+    "title": "Frontend",
+    "intro": "The core language of frontend is JS. From here, you can learn not only JS/CSS/HTML, but also React, NodeJS, Webpack as well. In the following lessons, you'll get a healthy understanding of each of front-end.",
+    "moreUrl": "#frontend"
+  },
+  {
+    "bannerUrl": "https://cosmos-x.oss-cn-hangzhou.aliyuncs.com/cardBanner3.svg",
+    "title": "Backend",
+    "intro": "The core language of frontend is JS. From here, you can learn not only JS/CSS/HTML, but also React, NodeJS, Webpack as well. In the following lessons, you'll get a healthy understanding of each of front-end.",
+    "moreUrl": "#backend"
+  },
+  {
+    "bannerUrl": "https://cosmos-x.oss-cn-hangzhou.aliyuncs.com/cardBanner4.svg",
+    "title": "Data & AI",
+    "intro": "The core language of frontend is JS. From here, you can learn not only JS/CSS/HTML, but also React, NodeJS, Webpack as well. In the following lessons, you'll get a healthy understanding of each of front-end.",
+    "moreUrl": "#data-ai"
+  }
+]
