+++ src/pages/Section3/PartIntro/index.module.css
@@ -3,7 +3,7 @@
   /* border: 1px solid #f00; */
   overflow: hidden;
   display: flex;
-  padding: 8rem 6rem;
+  padding: 5rem 6rem;
 }
 
 .titleFrom--left {
