+++ src/pages/index.module.css
@@ -51,19 +51,6 @@
   --uni-testimonial-text-color: var(--ifm-font-color-base);
 }
 
-.homepage__sections {
-  --ifm-hero-background-color: var(--dark-blue-grey);
-  background-image: url(https://cosmos-x.oss-cn-hangzhou.aliyuncs.com/homepage-cosmos-bg.svg),
-    linear-gradient(135deg, #273351 0, rgba(20, 20, 20) 100%);
-  background-repeat: repeat;
-  background-position: center;
-}
-
-.homepage__sectionsContainer {
-  max-width: 1880px;
-  margin: 0 auto;
-}
-
 html[data-theme='dark'] .section-bgwhite-block-right {
   background: var(--my-board--background);
 }
